Rename Form1 to MainForm
