+++ SoMRandomizer/Program.cs
@@ -35,7 +35,7 @@ namespace SoMRandomizer
             {
                 Application.EnableVisualStyles();
                 Application.SetCompatibleTextRenderingDefault(false);
-                Application.Run(new Form1());
+                Application.Run(new MainForm());
             }
             else
             {
